requirements.txt
annotated-types    0.6.0
blis               0.7.11
catalogue          2.0.10
certifi            2024.2.2
charset-normalizer 3.3.2
chatbot            1.5.2b0
chatterbot-corpus  1.2.0
ChatterBot2        1.1.0a7
click              8.1.7
cloudpathlib       0.16.0
colorama           0.4.6
confection         0.1.4
cymem              2.0.8
en-core-web-sm     3.7.1
idna               3.6
Jinja2             3.1.3
langcodes          3.3.0
MarkupSafe         2.1.5
mathparse          0.1.2
murmurhash         1.0.10
numpy              1.24.4
packaging          23.2
pip                24.0
preshed            3.0.9
pydantic           2.6.1
pydantic_core      2.16.2
python-dateutil    2.8.2
pytz               2024.1
PyYAML             3.12
requests           2.31.0
setuptools         69.0.3
six                1.16.0
smart-open         6.4.0
spacy              3.7.4
spacy-legacy       3.0.12
spacy-loggers      1.0.5
SQLAlchemy         1.3.24
srsly              2.4.8
thinc              8.2.3
tqdm               4.66.2
typer              0.9.0
typing_extensions  4.9.0
urllib3            2.2.1
wasabi             1.1.2
weasel             0.3.4
wheel              0.42.0
